+++ models/models.py
@@ -62,3 +62,40 @@ class ValueNetwork(nn.Module):
         x = F.relu(self.fc1(state))
         x = F.relu(self.fc2(x))
         return self.value(x)
+
+
+class PolicyNetwork(nn.Module):
+    """
+    Neural network for policy-based reinforcement learning.
+
+    Architecture:
+    - Input layer: Accepts `num_inputs` features representing the state.
+    - Hidden layers: Two fully connected layers with 256 units each and ReLU activation for non-linearity.
+    - Output layer: Produces `num_outputs`, representing action space size or logits.
+
+    Args:
+        num_inputs (int): Number of input features (state size).
+        num_outputs (int): Number of output features (action size).
+
+    Methods:
+        forward(x): Propagates the input through the network.
+    """
+
+    def __init__(self, num_inputs: int, num_outputs: int) -> None:
+        super().__init__()
+
+        self.net = nn.Sequential(
+            nn.Linear(num_inputs, 256), nn.ReLU(), nn.Linear(256, 256), nn.ReLU(), nn.Linear(256, num_outputs)
+        )
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        """
+        Forward pass of the network.
+
+        Args:
+            x (torch.Tensor): Input tensor of shape (batch_size, num_inputs).
+
+        Returns:
+            torch.Tensor: Output tensor of shape (batch_size, num_outputs).
+        """
+        return self.net(x)
