+++ pre_trained_models/sac_continuous_lunar_lander_training_2024-12-24_15-07-05/sac_checkpoint_best_metadata.json
@@ -1,0 +1,5 @@
+{
+    "episode": 351,
+    "steps_taken": 201009,
+    "latest_episode_reward": 321.0533471116506
+}
